renamed files and style changes for images
