+++ src/App.tsx
@@ -5,8 +5,8 @@ import * as ROUTES from './constants/routes';
 import { Store } from './store';
 import './App.css';
 import Employees from './components/Employees';
-import EmployeesGrid from './components/EmployeesGrid';
-import FeaturedEmployee from './components/FeaturedEmployee';
+import EmployeesGrid from './components/EmployeeTable';
+import FeaturedEmployee from './components/EmployeeGrid';
 
 // note to self: when you want to use React Router, the best way to start is with a navigation component that will
 // be used in the App component.
